Syntax to copy all features from a LHS annotation
Added the option within a standard RHS action to copy *all* features from an annotation in a LHS binding, rather than having to name each feature individually.  Syntax is:

:example.NewAnnotation = { :binding }
or
:example.NewAnnotation = { :binding.Type }

The logic is to extract the specified binding (and optionally subset it for the given type), take one of the bound annotations that has at least one feature, and copy all that annotation's features to the new annotation.  As with the existing { feature = :binding.Type.otherFeature } this is underspecified if there is more than one matching annotation in the binding.

You can freely mix and match this new construct with existing single-feature assignments and they will all be applied from left to right.  Thus

:example.NewAnnotation = { type = unknown, :binding }

treats "unknown" as the default value for type, but will override this if the annotation in :binding also has a type feature, whereas

:example.NewAnnotation = { :binding, type = unknown }

will always set type = unknown, even if the :binding annotation also has a type feature.

Closes #8